+++ client/jetbrains/webview/src/bridge-mock/index.ts
@@ -1,7 +1,7 @@
 import { SearchPatternType } from '@sourcegraph/shared/src/graphql-operations'
 
-import { Search } from '../search/App'
 import { Request } from '../search/jsToJavaBridgeUtil'
+import type { Search } from '../search/types'
 
 let savedSearch: Search = {
     query: 'r:github.com/sourcegraph/sourcegraph jetbrains',
